04 -  פרוייקטים אישים/אתר כתיבה עיוורת/אתר כתיבה עיוורת עם התחברות+הגנה/typing-tutor/Data_Base/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.1
